+++ src/tests/mass_sampling.cc
@@ -83,8 +83,8 @@ TEST(omega_decay) {
 
   printf("testing ρ⁰ distribution ...\n");
   hist_neutral.test(
-    [&](float m) {
-      float pcm = pCM(srts, mass_stable, m);
+    [&](double m) {
+      double pcm = pCM(srts, mass_stable, m);
       return type_rho_zero.spectral_function(m)
                     * pcm * blatt_weisskopf_sqr(pcm, 1); }
     //,"masses_rho_neutral.dat"
@@ -92,8 +92,8 @@ TEST(omega_decay) {
 
   printf("testing ρ⁺ distribution ...\n");
   hist_charged.test(
-    [&](float m) {
-      float pcm = pCM(srts, mass_stable, m);
+    [&](double m) {
+      double pcm = pCM(srts, mass_stable, m);
       return type_rho_plus.spectral_function(m)
                     * pcm * blatt_weisskopf_sqr(pcm, 1); }
     //,"masses_rho_charged.dat"
